remove reliance of typing on object identity, make types serializable
This makes types round-trippable to and from JSON. Internally, uses of
Set() replaced with sets encoded as arrays. This is not necessarily a
better design, but it is more consistent with the current term
representation in simple-eval, and thus reduces some friction in
dealing with terms that have embedded types, and, potentially, types
that refer back to terms.

Review by @IsaacFleetwood (no rush though...)